main_server/requirements.txt
asyncio~=3.4.3
mcp~=1.6.0
httpx~=0.28.1
requests~=2.32.3
pydantic~=2.11.3
fastapi~=0.115.0
fastapi-controllers~=0.3.0
uvicorn~=0.31.0
jsbeautifier~=1.15.4
pywebcopy~=7.0.2
setuptools
six
requests
lxml
cachecontrol
lxml_html_clean
filelock~=3.18.0
llama_cpp_python==0.2.63
openai==1.75.0
huggingface-hub==0.30.2
